feat(@vtmn/svelte): fix name progressbar
