+++ packages/showcases/svelte/stories/components/indicators/VtmnProgressbar/VtmnProgressbar.stories.svelte
@@ -1,82 +1,153 @@
 <script>
   import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
-  import { VtmnLoader } from '@vtmn/svelte';
-  import {
-    parameters,
-    argTypes,
-  } from '@vtmn/showcase-core/csf/components/indicators/loader.csf';
+  import { VtmnProgressbar } from '@vtmn/svelte';
+  import { parameters } from '@vtmn/showcase-core/csf/components/indicators/progressbar.csf';
   const img =
     'https://storage.googleapis.com/dkt-design-cdn/images/landscape-placeholder.jpg';
+  export const argTypes = {
+    size: {
+      type: { name: 'string', required: false },
+      description: 'The size of the loader.',
+      defaultValue: 'medium',
+      table: {
+        type: {
+          summary: 'string',
+        },
+        defaultValue: { summary: 'medium' },
+      },
+      control: {
+        type: 'radio',
+        options: ['small', 'medium', 'large'],
+      },
+    },
+    variant: {
+      type: { name: 'string', required: true },
+      description: 'Variant of the loader.',
+      defaultValue: 'linear',
+      control: {
+        type: 'radio',
+        options: ['linear', 'circular'],
+      },
+    },
+    progress: {
+      type: { name: 'number', required: false },
+      description: 'Value of the progressbar',
+      defaultValue: 30,
+      control: {
+        type: 'range',
+        min: 0,
+        max: 100,
+      },
+    },
+    label: {
+      type: { name: 'string', required: false },
+      description: 'Label displayed for the loader',
+      control: {
+        type: 'text',
+      },
+    },
+    indeterminate: {
+      type: { name: 'boolean', required: false },
+      description: 'Set the loader to indeterminate',
+      defaultValue: false,
+      control: {
+        type: 'boolean',
+      },
+    },
+    withTrack: {
+      type: { name: 'boolean', required: false },
+      description: 'Display track path',
+      defaultValue: false,
+      control: {
+        type: 'boolean',
+      },
+    },
+    img: {
+      type: { name: 'string', required: false },
+      description: 'Image source for circle loader',
+      defaultValue: '',
+      control: {
+        type: 'text',
+      },
+    },
+    imgAlt: {
+      type: { name: 'string', required: false },
+      description: 'Alternative text for the image.',
+      control: {
+        type: 'text',
+      },
+    },
+  };
 </script>
 
 <Meta
-  title="Components / Indicators / VtmnLoader"
-  component={VtmnLoader}
+  title="Components / Indicators / VtmnProgressbar"
+  component={VtmnProgressbar}
   {argTypes}
   {parameters}
 />
 
 <Template let:args>
-  <VtmnLoader {...args} />
+  <VtmnProgressbar {...args} />
 </Template>
 
 <Story name="Overview" />
 
 <Story name="Linear">
   <div class="block">
-    <VtmnLoader label="loading" progress="60" size="small" />
+    <VtmnProgressbar label="loading" progress="60" size="small" />
   </div>
   <div class="block">
-    <VtmnLoader progress="30" size="small" />
+    <VtmnProgressbar progress="30" size="small" />
   </div>
   <div class="block">
-    <VtmnLoader size="small" indeterminate />
+    <VtmnProgressbar size="small" indeterminate />
   </div>
 
   <div class="block">
-    <VtmnLoader label="loading" progress="25" size="medium" />
+    <VtmnProgressbar label="loading" progress="25" size="medium" />
   </div>
   <div class="block">
-    <VtmnLoader progress="70" size="medium" />
+    <VtmnProgressbar progress="70" size="medium" />
   </div>
   <div class="block">
-    <VtmnLoader size="medium" indeterminate />
+    <VtmnProgressbar size="medium" indeterminate />
   </div>
   <div class="block">
-    <VtmnLoader label="loading" progress="35" size="large" />
+    <VtmnProgressbar label="loading" progress="35" size="large" />
   </div>
   <div class="block">
-    <VtmnLoader progress="20" size="large" />
+    <VtmnProgressbar progress="20" size="large" />
   </div>
   <div class="block">
-    <VtmnLoader size="large" indeterminate />
+    <VtmnProgressbar size="large" indeterminate />
   </div>
 </Story>
 
 <Story name="Circular">
   <div class="story-loader-circular">
     <div>
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="10"
         variant="circular"
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="20"
         variant="circular"
         withTrack
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="30"
         variant="circular"
         progressIndicator
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="40"
@@ -84,14 +155,14 @@
         progressIndicator
         withTrack
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="50"
         variant="circular"
         {img}
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="60"
@@ -99,14 +170,14 @@
         withTrack
         {img}
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="70"
         variant="circular"
         indeterminate
       />
-      <VtmnLoader
+      <VtmnProgressbar
         class="story-loader-circular-content"
         size="small"
         progress="80"
@@ -117,27 +188,27 @@
     </div>
     <div class="story-loader-circular">
       <div>
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="10"
           variant="circular"
         />
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="20"
           variant="circular"
           withTrack
         />
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="30"
           variant="circular"
           progressIndicator
         />
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="40"
@@ -147,14 +218,14 @@
         />
       </div>
       <div>
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="50"
           variant="circular"
           {img}
         />
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="60"
@@ -162,14 +233,14 @@
           withTrack
           {img}
         />
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="70"
           variant="circular"
           indeterminate
         />
-        <VtmnLoader
+        <VtmnProgressbar
           class="story-loader-circular-content"
           size="medium"
           progress="80"
@@ -194,8 +265,16 @@
   :global(.story-loader-circular-content) {
     margin: 8px;
   }
-  :global(#anchor--components-indicators-vtmnloader--circular .docs-story),
-  :global(#anchor--components-indicators-vtmnloader--circular
+
+  :global(#anchor--components-indicators-vtmnprogressbar--linear .docs-story),
+  :global(#anchor--components-indicators-vtmnprogressbar--linear
+      .docs-story
+      > div:first-child) {
+    height: 450px;
+  }
+
+  :global(#anchor--components-indicators-vtmnprogressbar--circular .docs-story),
+  :global(#anchor--components-indicators-vtmnprogressbar--circular
       .docs-story
       > div:first-child) {
     height: 700px;
