+++ packages/sources/svelte/rollup.config.js
@@ -18,8 +18,8 @@ const src = {
       folder: 'indicators',
       components: [
         'VtmnBadge',
-        'VtmnLoader',
         'VtmnPrice',
+        'VtmnProgressbar',
         'VtmnRating',
         'VtmnTag',
       ],
