+++ packages/sources/svelte/src/components/index.js
@@ -7,8 +7,8 @@ export { default as VtmnTextInput } from './forms/VtmnTextInput/VtmnTextInput.sv
 
 // Indicators
 export { default as VtmnBadge } from './indicators/VtmnBadge/VtmnBadge.svelte';
-export { default as VtmnLoader } from './indicators/VtmnLoader/VtmnLoader.svelte';
 export { default as VtmnPrice } from './indicators/VtmnPrice/VtmnPrice.svelte';
+export { default as VtmnProgressbar } from './indicators/VtmnProgressbar/VtmnProgressbar.svelte';
 export { default as VtmnRating } from './indicators/VtmnRating/VtmnRating.svelte';
 export { default as VtmnTag } from './indicators/VtmnTag/VtmnTag.svelte';
