+++ packages/sources/svelte/src/components/indicators/VtmnProgressbar/VtmnProgressbar.svelte
@@ -1,20 +1,21 @@
 <script>
   import { VTMN_LOADER_VARIANT, VTMN_LOADER_SIZE } from './enums';
-  import VtmnLoaderCircular from './VtmnLoaderCircular.svelte';
-  import VtmnLoaderLinear from './VtmnLoaderLinear.svelte';
+  import VtmnProgressbarCircular from './VtmnProgressbarCircular.svelte';
+  import VtmnProgressbarLinear from './VtmnProgressbarLinear.svelte';
 
   export let variant = VTMN_LOADER_VARIANT.LINEAR;
+  export let size = VTMN_LOADER_SIZE.MEDIUM;
   export let label = '';
   export let progress = 0;
   export let indeterminate = false;
-  export let size = VTMN_LOADER_SIZE.MEDIUM;
   export let progressIndicator = false;
   export let withTrack = false;
   export let img = '';
+  export let imgAlt;
 </script>
 
 {#if variant === VTMN_LOADER_VARIANT.LINEAR}
-  <VtmnLoaderLinear
+  <VtmnProgressbarLinear
     {label}
     {progress}
     {indeterminate}
@@ -23,17 +24,18 @@
     {...$$restProps}
   />
 {:else if variant === VTMN_LOADER_VARIANT.CIRCULAR}
-  <VtmnLoaderCircular
+  <VtmnProgressbarCircular
     {progress}
     {indeterminate}
     {size}
     {progressIndicator}
     {withTrack}
     {img}
+    {imgAlt}
     {...$$restProps}
   />
 {/if}
 
 <style>
-  @import '@vtmn/css-loader';
+  @import '@vtmn/css-progressbar';
 </style>
