+++ packages/sources/svelte/src/components/indicators/VtmnProgressbar/VtmnProgressbarLinear.svelte
