+++ packages/sources/svelte/src/components/indicators/VtmnProgressbar/enums.js
